+++ src/main/java/dev/_2lstudios/squidgame/commands/SquidGameCommand.java
@@ -10,6 +10,7 @@ import dev._2lstudios.squidgame.commands.admin.SquidEditArenaCommand;
 import dev._2lstudios.squidgame.commands.admin.SquidWandCommand;
 import dev._2lstudios.squidgame.commands.game.SquidJoinCommand;
 import dev._2lstudios.squidgame.commands.game.SquidLeaveCommand;
+import dev._2lstudios.squidgame.commands.game.SquidStartCommand;
 
 @Command(name = "squidgame")
 public class SquidGameCommand extends CommandListener {
@@ -23,6 +24,7 @@ public class SquidGameCommand extends CommandListener {
         // Game commands
         this.addSubcommand(new SquidLeaveCommand());
         this.addSubcommand(new SquidJoinCommand());
+        this.addSubcommand(new SquidStartCommand());
     }
 
     @Override