+++ CHANGELOG.md
@@ -11,7 +11,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Now tutorial title is configurable per tutorial item, like subtitle.
 - Added configurable sounds for red-light and green-light game.
-- Added reward system with commands executed by console.
+- Added reward system with commands executed by console..
+
+### Fixed
+
+- Fixed /squid leave command throw exception when arena is starting
 
 ## [0.1.2] - 2021-10-16